refactor: s3 api, hook 분리
