+++ src/services/s3/apis/uploadImageToS3.ts
@@ -1,0 +1,23 @@
+import type { PreparedUrls } from '~/services/s3/utils';
+
+import { publicAxios } from '~/utils/axios';
+
+export interface UploadImageToS3Params {
+  url: string;
+  file: File;
+}
+
+export type UploadImageToS3ApiData = PreparedUrls;
+
+export const uploadImageToS3 = (params: UploadImageToS3Params) => {
+  const { url, file } = params;
+
+  return publicAxios
+    .put<UploadImageToS3ApiData>(url, file, {
+      headers: {
+        'Content-Type': 'image/webp',
+      },
+      baseURL: '',
+    })
+    .then((res) => res.data);
+};
