+++ src/services/s3/index.ts
@@ -1,2 +1,3 @@
 export * from './utils';
 export * from './apis';
+export * from './hooks';
